+++ src/subset/possessives/index.js
@@ -1,20 +1,21 @@
 'use strict';
 const Text = require('../../text');
-const Terms = require('../../paths').Terms;
+// const Terms = require('../../paths').Terms;
 
 const methods = {
-  //remove the 's on it
-  normalize() {
+  //remove the 's on the end of the word
+  strip() {
     this.list.forEach((ts) => {
       let t = ts.terms[ts.terms.length - 1];
-      console.log(t.text);
+      t.text = t.text.replace(/'s$/, '');
     });
+    return this;
   }
 };
 
 const find = function(r, n) {
   r = r.match('#Possessive+');
-  r.debug();
+  r = r.splitAfter('#Comma');
   if (typeof n === 'number') {
     r = r.get(n);
   }