+++ changelog.md
@@ -44,6 +44,8 @@ compromise uses semver, and pushes to npm frequently
 - expose internal `sentences().isQuestion()` method
 - `.join()` as an alias for `.flatten()`
 - slightly different behavior for wildcards in capture-groups [pull/472](https://github.com/spencermountain/compromise/pull/472)
+- `.possessives()` subset + `#Possessive` tagging fixes
+- hide massive `world` output for console.log of a term
 
 ##### 11.0.0
 - support for persistent lexicon/tagset changes